pyproject.toml
[project]
name = "gpu-cluster-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
